refactor: extract aggregation selector
